+++ IntroToEF.Data/Repositories/IMichielRepo.cs
@@ -3,19 +3,34 @@ using System.Collections.Generic;
 
 namespace IntroToEF.Data.Repositories
 {
-    public interface ISamuraiRepo
+    public interface IMichielRepo
     {
         void AddDifferentObjectsToContext();
+
         void AddSamurai(Samurai samurai);
+
         void AddSamurai(string name);
+
         void AddSamurais(List<Samurai> samurais);
+
         void DeleteSamurai(int id);
-        Samurai GetSamurai(int id, bool fetchAllRelatedData = false);
-        Samurai GetSamuraiWithIncludedData(int id);
-        List<Samurai> GetSamurais();
-        void UpdateSamurai(Samurai samurai);
-        void UpdateSamurais();
-        Samurai GetSamuraiWhereNameContains(string text);
+
         List<Samurai> GetResultFromStoredProcedure(string text);
+
+        Samurai GetSamurai(int id, bool fetchAllRelatedData = false);
+
+        Samurai GetSamuraiByName(string name);
+
+        List<Samurai> GetSamurais();
+
+        List<Samurai> GetSamuraisByName(string name);
+
+        Samurai GetSamuraiWithIncludedData(int id);
+
+        void GetSamuraiWithSql();
+
+        void UpdateSamurai(Samurai samurai);
+
+        void UpdateSamurais();
     }
 }
