+++ IntroToEF.Data/SamuraiAreCoolContext.cs
@@ -5,10 +5,11 @@ using System;
 namespace IntroToEF.Data
 {
     // Context is absolutely essential in EF -> MUST inherit from DBContext
-    public class SamuraiContext : DbContext
+    public class SamuraiAreCoolContext : DbContext
     {
         // Each entity that needs a table needs to be defined here
         public DbSet<Quote> Quotes { get; set; }
+
         public DbSet<Samurai> Samurais { get; set; }
         public DbSet<Horse> Horses { get; set; }
         public DbSet<Battle> Battles { get; set; }
@@ -18,8 +19,8 @@ namespace IntroToEF.Data
         // Override the OnConfigure to dictate which database is being used and the type of said DB
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.UseSqlServer(CONNECTION)
-                .LogTo(Console.WriteLine);
+            optionsBuilder.UseSqlServer(CONNECTION);
+            //.LogTo(Console.WriteLine); (om te zien wat hij exact in de database doet in uw console)
         }
     }
 }
