+++ IntroToEF/Menu.cs
@@ -21,7 +21,6 @@ namespace IntroToEF
 
         public void MainMenu()
         {
-            nav.ShowKatana();
             nav.MenuTopBanner("Add data", "Select Samurai", "Search Data");
 
             switch (nav.MenuOptions(3))
@@ -82,7 +81,7 @@ namespace IntroToEF
         public void SelectSamurai()
         {
             nav.MenuTopBanner("Update Samurai", "Delete Samurai");
-            logic.ShowSamurai();
+            logic.GetAllSamurais();
             switch (nav.MenuOptions(2))
             {
                 case 1:
